no encuentro falla cartContext aun

+++ client/src/Context/AuthContext.js
@@ -2,8 +2,8 @@ import { createContext, useState } from 'react';
 import { useNavigate } from 'react-router-dom';
 import axios from 'axios';
 
-const API_LOGIN = 'http://localhost:8080/login';
-// const API_REGISTER = 'http://localhost:8080/register';
+const API_LOGIN = 'http://localhost:8080/api/sign/login';
+const API_REGISTER = 'http://localhost:8080/sign/register';
 const API_LOGOUT = 'http://localhost:8080/logout';
 
 
@@ -14,19 +14,27 @@ export const AuthContext = createContext()
 export const AuthProvider = ({ children }) => {
   const [ auth, setAuth ] = useState(null)
   const [ user, setUser ] = useState(null)
-  const [ userCredentials, setUserCredentials ] = useState({
+  const [ loginCredentials, setLoginCredentials ] = useState({
     email: '',
     password: ''
+  })
+  const [ registerCredentials, setRegisterCredentials ] = useState({
+    firstName: '',
+    lastName: '',
+    userName: '',
+    phone: '',
+    email: '',
+    password: '',
+    confirmPassword: ''
   })
 
   const navigate = useNavigate();
 
   const handleSubmitLogin = (e) => {
     e.preventDefault();
-    axios.post(API_LOGIN, userCredentials)
+    axios.post(API_LOGIN, loginCredentials)
       .then((res) => {
         const token = res.data;
-        console.log("TOKENFRONT: ", res.data.user);
         setAuth(token)
         setUser(res.data.user)
         localStorage.setItem('user', res.data)
@@ -39,10 +47,25 @@ export const AuthProvider = ({ children }) => {
 
   const handleChange = (e) => {
     e.preventDefault();
-    setUserCredentials({
-      ...userCredentials,
+    setLoginCredentials({
+      ...loginCredentials,
       [e.target.name]: e.target.value
     })
+  }
+
+  const handleSubmitRegister = (e) => {
+    e.preventDefault();
+    axios.post(API_REGISTER, registerCredentials)
+      .then((res) => {
+        const token = res.data;
+        setAuth(token)
+        setUser(res.data.user)
+        localStorage.setItem('user', res.data)
+        navigate('/store')
+      })
+      .catch((err) => {
+        console.log(err)
+      })
   }
 
   const handleLogout = () => {
@@ -55,9 +78,12 @@ export const AuthProvider = ({ children }) => {
       auth: auth,
       user: user,
       setAuth: setAuth,
-      userCredentials: userCredentials,
-      setUserCredentials: setUserCredentials,
+      loginCredentials: loginCredentials,
+      setLoginCredentials: setLoginCredentials,
+      registerCredentials: registerCredentials,
+      setRegisterCredentials: setRegisterCredentials,
       handleSubmitLogin: handleSubmitLogin,
+      habdleSubmitRegister: handleSubmitRegister,
       handleChange: handleChange,
       handleLogout: handleLogout,
   }
